Add files via upload
This is my first project in the exciting world of microcontrollers and embedded computing. This code runs the same test that the Arduino development team uses to verify the functionality of the board before it is shipped out. It is a simple program that blinks the built in LED connected to pin 13 by turning it on and off every 1 second. Demonstration is on my YouTube channel.